refactor: Move funtions to semantically correct package
