+++ pkg/namespace/namespace.go
@@ -15,7 +15,11 @@
 package namespace
 
 import (
+	"fmt"
+	"os"
+	"path/filepath"
 	"sort"
+	"syscall"
 
 	"github.com/parca-dev/parca-agent/pkg/cgroup"
 	"github.com/prometheus/procfs"
@@ -37,6 +41,7 @@ func (n ns) Inode() uint32 {
 	return n.Namespace.Inode
 }
 
+// NamespacesForPID returns the namespaces of the process with the given PID.
 func NamespacesForPID(pid int) ([]Namespace, error) {
 	proc, err := procfs.NewProc(pid)
 	if err != nil {
@@ -115,4 +120,17 @@ func AdjacentPIDs(pid int) ([]int, error) {
 	}
 	sort.Ints(pids)
 	return pids, nil
+}
+
+// MountNamespaceInode returns the inode of the mount namespace of the given pid.
+func MountNamespaceInode(pid int) (uint64, error) {
+	fileinfo, err := os.Stat(filepath.Join("/proc", fmt.Sprintf("%d", pid), "ns/mnt"))
+	if err != nil {
+		return 0, err
+	}
+	stat, ok := fileinfo.Sys().(*syscall.Stat_t)
+	if !ok {
+		return 0, fmt.Errorf("not a syscall.Stat_t")
+	}
+	return stat.Ino, nil
 }
